docs(tutorial): aspect-correction lives in step 2; bigger equilateral throughout
Reordered:
  Step 2 = "A second uniform" (was step 3) — introduce `resolution`
           and aspect-correction here.
  Step 3 = "Compose from the catalog" (was step 2) — the noise overlay
           now follows the aspect-ratio fix and inherits it.

Bigger, dead-centered equilateral triangle baked everywhere positions
appear: (-0.7, -0.4), (0.7, -0.4), (0, 0.8). Centroid at exactly
(0, 0); side lengths 1.4 / 1.389 / 1.389 — visually equilateral within
1%, fills more of the canvas than the previous (-0.6, -0.35, 0.69)
shape.

Aspect-correction now lives in every shader from step 2 onward, so the
"nice triangle" promise holds for every subsequent demo:
  - Step 2: resolution-uniform inline shader (was step 3 before swap).
  - Step 3: compose-noise inline shader picks up the resolution
    uniform too; UV coords still come from the pre-correction
    positions so the noise field stays stable.
  - Step 4: vertex-gradient inline shader takes resolution and
    aspect-corrects per-vertex.
  - Step 5: particle-trail inline shader aspect-corrects each
    particle's final NDC position so the field stays evenly
    distributed and per-particle triangles stay equilateral.
  - Step 6: skipped — particles render into a square 1024x1024
    intermediate, postfx is fullscreen, neither needs correction.

Step 1's URL'd `triangle.wgsl` deliberately keeps no resolution
uniform — the reader must see the squish there before step 2 motivates
the fix.

verify_tutorial_shaders updated to reflect the renumbering and the new
binding layouts; multipass smoke test still passes (262144-byte
readback).
