+++ docs/tutorials/01-hello-triangle/06-postfx-pass/example.js
@@ -103,9 +103,9 @@ export async function setup(renderer, _target) {
 
     const mesh = new Mesh();
     mesh.addVertices([
-        new Vertex([-0.6, -0.5, 0.0]).set("color", [0.95, 0.30, 0.42]),
-        new Vertex([ 0.6, -0.5, 0.0]).set("color", [0.30, 0.85, 0.55]),
-        new Vertex([ 0.0,  0.7, 0.0]).set("color", [0.30, 0.55, 0.95]),
+        new Vertex([-0.7, -0.4, 0.0]).set("color", [0.95, 0.30, 0.42]),
+        new Vertex([ 0.7, -0.4, 0.0]).set("color", [0.30, 0.85, 0.55]),
+        new Vertex([ 0.0,  0.8, 0.0]).set("color", [0.30, 0.55, 0.95]),
     ]);
 
     const TAU = Math.PI * 2;
