+++ docs/website/public/triangle.wgsl
@@ -10,12 +10,11 @@ struct VOut { @builtin(position) pos: vec4<f32> };
 
 @vertex
 fn vs_main(@builtin(vertex_index) i: u32) -> VOut {
-    // Equilateral triangle in NDC. apex_y / base_y come from
-    // height = side * sqrt(3) / 2 with side = 1.2.
+    // Equilateral triangle in NDC, centroid at origin.
     var p = array<vec2<f32>, 3>(
-        vec2<f32>(-0.6, -0.35),
-        vec2<f32>( 0.6, -0.35),
-        vec2<f32>( 0.0,  0.69),
+        vec2<f32>(-0.7, -0.4),
+        vec2<f32>( 0.7, -0.4),
+        vec2<f32>( 0.0,  0.8),
     );
     var out: VOut;
     out.pos = vec4<f32>(p[i], 0.0, 1.0);
